requirements.txt
python-dotenv
advent-of-code-data
rich
